# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shop API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Регистрация",
                "operationId": "651f8e74dc32f06cb7c61fdca24605e5",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "surname",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "address",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "password_confirmed",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "number",
                                                "example": "1"
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Андрей"
                                            },
                                            "surname": {
                                                "type": "string",
                                                "example": "Лясковский"
                                            },
                                            "address": {
                                                "type": "string",
                                                "example": "1-й Красногвардейский проезд, 22с1, Москва"
                                            },
                                            "phone": {
                                                "type": "string",
                                                "example": "+7 (903) 111-11-11"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "lae3145@mail.ru"
                                            },
                                            "email_verified_at": {
                                                "type": "string",
                                                "example": "null"
                                            },
                                            "created_at": {
                                                "type": "time",
                                                "example": "2023-07-06T08:27:30.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "time",
                                                "example": "2023-07-06T09:45:07.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/api/add_category": {
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Добавить категорию",
                "operationId": "3bfa0b3bbb02b944cc93a2c250433afe",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category successfully created"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/delete_category": {
            "delete": {
                "tags": [
                    "Category"
                ],
                "summary": "Удалить категорию",
                "operationId": "ceefbcc8c0f0d5fe6b3792ad2909be09",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category successfully deleted"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/favorites": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Избранное",
                "operationId": "76f556b3d2e972ad673f3a445cc18e6a",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "number",
                                                "example": "1"
                                            },
                                            "product_id": {
                                                "type": "number",
                                                "example": "2"
                                            },
                                            "user_id": {
                                                "type": "number",
                                                "example": "5"
                                            },
                                            "created_at": {
                                                "type": "time",
                                                "example": "2023-07-06T08:27:30.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "time",
                                                "example": "2023-07-06T09:45:07.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/add_favorites": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Добавить в избранное",
                "operationId": "60fe14d188fcac564c22b5eab4423601",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "number",
                                                "example": "1"
                                            },
                                            "product_id": {
                                                "type": "number",
                                                "example": "2"
                                            },
                                            "user_id": {
                                                "type": "number",
                                                "example": "5"
                                            },
                                            "created_at": {
                                                "type": "time",
                                                "example": "2023-07-06T08:27:30.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "time",
                                                "example": "2023-07-06T09:45:07.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/delete_favorites": {
            "delete": {
                "tags": [
                    "Profile"
                ],
                "summary": "Удалить товар из избранного",
                "operationId": "451b84ac8d2b2d6ee58ce7212f48d8c1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Favorite successfully deleted"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/basket": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Корзина",
                "operationId": "149d7ae0f454ce951afaabc22799157d",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "number",
                                                "example": "1"
                                            },
                                            "product_id": {
                                                "type": "number",
                                                "example": "2"
                                            },
                                            "user_id": {
                                                "type": "number",
                                                "example": "5"
                                            },
                                            "created_at": {
                                                "type": "time",
                                                "example": "2023-07-06T08:27:30.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "time",
                                                "example": "2023-07-06T09:45:07.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/add_basket": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Добавить в корзину",
                "operationId": "07b0aa9821b01274791153c69cb997a5",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "number",
                                                "example": "1"
                                            },
                                            "product_id": {
                                                "type": "number",
                                                "example": "2"
                                            },
                                            "user_id": {
                                                "type": "number",
                                                "example": "5"
                                            },
                                            "created_at": {
                                                "type": "time",
                                                "example": "2023-07-06T08:27:30.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "time",
                                                "example": "2023-07-06T09:45:07.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/delete_basket": {
            "delete": {
                "tags": [
                    "Profile"
                ],
                "summary": "Удалить товар из коризны",
                "operationId": "06e594a2bf19b6afe387847ec5f0f0ee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Basket successfully deleted"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "description": "Your token",
                "name": "Auth",
                "in": "header"
            }
        }
    }
}